-Added a controller superclass for methods that are common to both SolveController and PlayController class -Fixed bug causing window size to change -Fixed bug in checkInput that returns false if any 2 nos entered -checkInput now checks duplicates in rows, columns and 3x3 grids -Fixed padding issue for sudoku grid -Reduced code size for Solver class (checkCandidate implementation changed from if else mess to for loops)
