+++ src/Sudoku/GUI/Resources/MainView.css
